внесла правки после ревью
