+++ ImageFeed/Auth/CleanService.swift
@@ -1,0 +1,19 @@
+//
+//  CleanService.swift
+//  ImageFeed
+//
+//  Created by Irina Golubinskaya on 10.01.2024.
+//
+
+import WebKit
+
+final class CleanService {
+    static func clean() {
+        HTTPCookieStorage.shared.removeCookies(since: Date.distantPast)
+        WKWebsiteDataStore.default().fetchDataRecords(ofTypes: WKWebsiteDataStore.allWebsiteDataTypes()) { records in
+            records.forEach { record in
+                WKWebsiteDataStore.default().removeData(ofTypes: record.dataTypes, for: [record], completionHandler: {})
+            }
+        }
+    }
+}
