+++ ImageFeed/Models/ProfileImage.swift
@@ -1,0 +1,14 @@
+//
+//  ProfileImage.swift
+//  ImageFeed
+//
+//  Created by Irina Golubinskaya on 10.01.2024.
+//
+
+import Foundation
+
+struct ProfileImage: Codable {
+    let small: String?
+    let medium: String?
+    let large: String?
+}
