+++ ImageFeed/Models/ProfileResult.swift
@@ -1,0 +1,24 @@
+//
+//  ProfileResult.swift
+//  ImageFeed
+//
+//  Created by Irina Golubinskaya on 10.01.2024.
+//
+
+import Foundation
+
+struct ProfileResult: Decodable {
+    let username: String
+    let firstName: String?
+    let lastName: String?
+    let bio: String?
+    let profileImage: ProfileImage?
+    
+    private enum CodingKeys: String, CodingKey {
+        case username
+        case firstName = "first_name"
+        case lastName = "last_name"
+        case bio
+        case profileImage = "profile_image"
+    }
+}
