+++ ImageFeed/Models/UrlsResult.swift
@@ -1,0 +1,16 @@
+//
+//  UrlsResult.swift
+//  ImageFeed
+//
+//  Created by Irina Golubinskaya on 10.01.2024.
+//
+
+import Foundation
+
+struct UrlsResult: Codable {
+    let raw: String?
+    let full: String?
+    let regular: String?
+    let small: String?
+    let thumb: String?
+}
