+++ ImageFeed/Models/UserResult.swift
@@ -1,0 +1,12 @@
+//
+//  UserResult.swift
+//  ImageFeed
+//
+//  Created by Irina Golubinskaya on 10.01.2024.
+//
+
+import Foundation
+
+struct UserResult: Codable {
+    let profileImage: ProfileImage?
+}
