+++ ImageFeed/Profile/ProfileImageService.swift
@@ -62,13 +62,3 @@ final class ProfileImageService {
         return request
     }
 }
-
-struct UserResult: Codable {
-    let profileImage: ProfileImage?
-}
-
-struct ProfileImage: Codable {
-    let small: String?
-    let medium: String?
-    let large: String?
-}
