+++ ImageFeed/Profile/ProfileService.swift
@@ -61,61 +61,3 @@ final class ProfileService {
         task.resume()
     }
 }
-
-struct ProfileResult: Decodable {
-    let username: String
-    let firstName: String?
-    let lastName: String?
-    let bio: String?
-    let profileImage: ProfileImage?
-    
-    private enum CodingKeys: String, CodingKey {
-        case username
-        case firstName = "first_name"
-        case lastName = "last_name"
-        case bio
-        case profileImage = "profile_image"
-    }
-}
-
-struct Links: Decodable {
-    let selfLink: String
-    let html: String
-    let photos: String
-    let likes: String
-    let portfolio: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case selfLink = "self"
-        case html
-        case photos
-        case likes
-        case portfolio
-    }
-}
-
-struct Social: Decodable {
-    let instagramUsername: String
-    let portfolioURL: String
-    let twitterUsername: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case instagramUsername = "instagram_username"
-        case portfolioURL = "portfolio_url"
-        case twitterUsername = "twitter_username"
-    }
-}
-
-struct Badge: Decodable {
-    let title: String
-    let primary: Bool
-    let slug: String
-    let link: String
-    
-    private enum CodingKeys: String, CodingKey {
-        case title
-        case primary
-        case slug
-        case link
-    }
-}
